python_app/requirements.txt
prometheus_client
Flask
Werkzeug
PyJWT
pika